chore(jans-cedarling): refactor, move `log entry` to log module
