+++ jans-cedarling/cedarling/src/log/test.rs
@@ -8,13 +8,7 @@ use std::{
 };
 
 use super::*;
-use crate::{
-    log::stdout_logger::TestWriter,
-    models::{
-        app_types,
-        log_entry::{LogEntry, LogType},
-    },
-};
+use crate::{log::stdout_logger::TestWriter, models::app_types};
 use interface::LogWriter;
 use nop_logger::NopLogger;
 use stdout_logger::StdOutLogger;
